run formatter

+++ frontend/src/app/team-management/member-list/member-list-table/member-list-table.component.ts
@@ -1,26 +1,26 @@
-import { Component, Input, OnDestroy, OnInit } from '@angular/core';
-import { MatTableDataSource } from '@angular/material/table';
-import { UserTableEntry } from '../../../shared/types/model/UserTableEntry';
-import { getFullNameFromUser, User } from '../../../shared/types/model/User';
-import { Team } from '../../../shared/types/model/Team';
-import { TeamService } from '../../../services/team.service';
-import { UserService } from '../../../services/user.service';
-import { getRouteToUserDetails } from '../../../shared/routeUtils';
-import { BehaviorSubject, filter, mergeMap, Subject, takeUntil } from 'rxjs';
-import { UserTeam } from '../../../shared/types/model/UserTeam';
-import { DialogService } from '../../../services/dialog.service';
+import { Component, Input, OnDestroy, OnInit } from "@angular/core";
+import { MatTableDataSource } from "@angular/material/table";
+import { UserTableEntry } from "../../../shared/types/model/UserTableEntry";
+import { User } from "../../../shared/types/model/User";
+import { Team } from "../../../shared/types/model/Team";
+import { TeamService } from "../../../services/team.service";
+import { UserService } from "../../../services/user.service";
+import { getRouteToUserDetails } from "../../../shared/routeUtils";
+import { BehaviorSubject, filter, mergeMap, Subject, takeUntil } from "rxjs";
+import { UserTeam } from "../../../shared/types/model/UserTeam";
+import { DialogService } from "../../../services/dialog.service";
 
 @Component({
-  selector: 'app-member-list-table',
-  templateUrl: './member-list-table.component.html',
-  styleUrl: './member-list-table.component.scss',
+  selector: "app-member-list-table",
+  templateUrl: "./member-list-table.component.html",
+  styleUrl: "./member-list-table.component.scss",
 })
 export class MemberListTableComponent implements OnInit, OnDestroy {
   @Input({ required: true }) dataSource!: MatTableDataSource<UserTableEntry>;
   @Input({ required: true }) selectedTeam$!: BehaviorSubject<undefined | Team>;
 
-  private allColumns = ['icon', 'name', 'roles', 'teams', 'okr_champion'];
-  private teamColumns = ['icon', 'name', 'role'];
+  private allColumns = ["icon", "name", "roles", "teams", "okr_champion"];
+  private teamColumns = ["icon", "name", "role"];
   private unsubscribe$ = new Subject<void>();
 
   displayedColumns: string[] = this.allColumns;
@@ -32,9 +32,10 @@ export class MemberListTableComponent implements OnInit, OnDestroy {
   ) {}
 
   ngOnInit() {
-    this.selectedTeam$.pipe(takeUntil(this.unsubscribe$)).subscribe((team) => {
-      team ? this.setColumnForTeam(team) : this.setColumnsForAllTeams();
-    });
+    this.selectedTeam$.pipe(takeUntil(this.unsubscribe$))
+      .subscribe((team) => {
+        team ? this.setColumnForTeam(team) : this.setColumnsForAllTeams();
+      });
   }
 
   ngOnDestroy() {
@@ -45,7 +46,7 @@ export class MemberListTableComponent implements OnInit, OnDestroy {
   private setColumnForTeam(team: Team) {
     this.displayedColumns = [...this.teamColumns];
     if (team.writeable) {
-      this.displayedColumns.push('menu');
+      this.displayedColumns.push("menu");
     }
   }
 
@@ -61,15 +62,15 @@ export class MemberListTableComponent implements OnInit, OnDestroy {
       team: this.selectedTeam$.value?.name,
     };
     this.dialogService
-      .openConfirmDialog('CONFIRMATION.DELETE.USER_FROM_TEAM', i18nData)
+      .openConfirmDialog("CONFIRMATION.DELETE.USER_FROM_TEAM", i18nData)
       .afterClosed()
       .pipe(
-        filter((confirm) => confirm),
-        mergeMap(() => this.teamService.removeUserFromTeam(entry.id, this.selectedTeam$.value!)),
+        filter((confirm) => confirm), mergeMap(() => this.teamService.removeUserFromTeam(entry.id, this.selectedTeam$.value!)),
       )
       .subscribe(() => {
         this.userService.reloadUsers();
-        this.userService.reloadCurrentUser().subscribe();
+        this.userService.reloadCurrentUser()
+          .subscribe();
       });
   }
 
@@ -77,11 +78,13 @@ export class MemberListTableComponent implements OnInit, OnDestroy {
     // make a copy and set value only after successful request
     const newUserTeam = { ...userTeam };
     newUserTeam.isTeamAdmin = isAdmin;
-    this.teamService.updateOrAddTeamMembership(userTableEntry.id, newUserTeam).subscribe(() => {
-      userTeam.isTeamAdmin = isAdmin;
-      this.userService.reloadUsers();
-      this.userService.reloadCurrentUser().subscribe();
-    });
+    this.teamService.updateOrAddTeamMembership(userTableEntry.id, newUserTeam)
+      .subscribe(() => {
+        userTeam.isTeamAdmin = isAdmin;
+        this.userService.reloadUsers();
+        this.userService.reloadCurrentUser()
+          .subscribe();
+      });
   }
 
   getMemberDetailsLink(user: User, team?: Team) {
@@ -92,7 +95,7 @@ export class MemberListTableComponent implements OnInit, OnDestroy {
   // otherwise the method will throw an exception
   getSingleUserTeam(userTableEntry: UserTableEntry): UserTeam {
     if (userTableEntry.userTeamList.length !== 1) {
-      throw Error('it should have exactly one UserTeam at this point');
+      throw Error("it should have exactly one UserTeam at this point");
     }
     return userTableEntry.userTeamList[0];
   }
